fastlane from IzzyOnDroid & integrating metadata from Triple-T (#513)
this PR provides you with a "fastlane starter package" (as discussed in
#476) – to give it into your hands to define how your app is presented,
and make it easy to keep descriptions & graphics in sync with your
development. Some notes to sum them up in a place easy to find for you:

* the [IzzyOnDroid Fastlane
Documentation](https://gitlab.com/IzzyOnDroid/repo/-/wikis/Fastlane) can
guide you with maintaining the structures here
* take special care for limits, e.g. per-release changelogs can have max
500 chars, `full_description.txt` max 4,000 chars, graphics must have
specific aspect ratios etc.
* you can have "bigger graphics" here; the IoD updater takes care to
resize/optimize them as needed
* for the `full_description.txt` here I've used HTML compressed into a
single line, to prevent the fdroid software converting each line break
to a `<br>`. The tags used for this are supported by F-Droid.org as
well, and to my knowledge even by PlayStore.
* whenever you add a new "type" of metadata that has not been there
before (e.g. the first changelog, a featureGraphic if it wasn't yet
present), please let us know so we can check and enable it.
* Details on all this in the linked documentation.
* Once merged, your changes will be synced whenever a new release is
pulled in here – to make sure it stays up-to-date with the app itself.
* Should you have questions, be welcome to ask :wink:

And now: Enjoy!

---------

Co-authored-by: TrianguloY <correo--correo@hotmail.com>